+++ lib/oli_web.ex
@@ -35,6 +35,8 @@ defmodule OliWeb do
         root: "lib/oli_web/templates",
         namespace: OliWeb
 
+      use Appsignal.Phoenix.View
+
       # Import convenience functions from controllers
       import Phoenix.Controller,
         only: [get_flash: 1, get_flash: 2, view_module: 1, view_template: 1]