+++ lib/oli_web/endpoint.ex
@@ -1,5 +1,6 @@
 defmodule OliWeb.Endpoint do
   use Phoenix.Endpoint, otp_app: :oli
+  use Appsignal.Phoenix
 
   @session_options [
     store: :cookie,