+++ CHANGELOG.md
@@ -4,12 +4,26 @@
 
 ### Bug Fixes
 
+### Enhancements
+
+## 0.14.2 (2021-10-28)
+
+### Bug Fixes
+
+- Make score display in gradebook more robust
+
+## 0.14.1 (2021-10-28)
+
+### Bug Fixes
+
 - Fix an issue with Apply Update button and move content updates to async worker
 - Fix text editor issue where image resizing didn't work from right drag handle
 - Fix text editor issue where link editing tooltip could overlap with formatting toolbar
 - Fix the "insert content" toolbar positioning
 - Fix an issue where the button for inserting a table would move around unless the mouse was positioned in just the right way
 - Fix an issue where previewing a project with no pages crashes
+- Fix some issues related to searching and viewing sortable tables
+- Fix an issue where activity submissions would not display activity feedback
 
 ### Enhancements
 
@@ -17,6 +31,8 @@
 - New Popup page element
 - New Instructor Course Section overview page
 - Allow LMS and System admins to unlink LMS section
+- Gradebook and graded page per student details with instructor centered grade override
+- Student specific progress display
 
 ## 0.14.0 (2021-10-13)